STYLE: pin store 이름 변경
